Created MongoDB setup connection & reintroduced axios
Created an initial connection to MongoDB. Created a new directory
in backEnd/ for database logic and created a dbConnection.js file
to setup the Mongo database. Reintroduced the axios package in
backEnd/routes/index.js and tested out the app and it works as
intended. Edited .gitignore to ignore the venv folder in the backend.