App matches user input with JSON file entries
The app accepts user input and matched it with JSON by law entries
and properly outputs the related entries from the JSON file. A new
python file was created for the logic to work. callToJSON.py calls
the JSON file and api.py has the code for the output format. The
app works as intended