Added code for Submit button toast pop up
A new function was added that would trigger a toast message to
show up when the user interacts with the submit button in the
2nd page of the reportViolation process. Also, the input fields
text color was changed to black to make the user input values
show up. Previously it was white text so it would be invisible
in a white background. The code now works as intended.