Added the feature to upload images and open camera
Created a feature to allow the user to upload images (required by
the app to upload evidence) and to allow the user to have access
to the camera of their mobile device. The logic for those work
as intended. The UI for Page 2 of report violatio was also edited
to allow the user to deselect uploaded images as well as make that
interface scrollable to view 3+ uploaded images.